requirements.txt
Jinja2
boto3
rich
python-dotenv
click